requirements.txt
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
psycopg[binary]==3.2.6
python-dotenv==1.0.1
reportlab==4.2.5
